add calibration-tool
