+++ software/shepherd-calibrate/setup.cfg
@@ -1,0 +1,54 @@
+[metadata]
+name = shepherd_cal
+version = attr: shepherd_cal.__version__
+author = Kai Geissdoerfer, Ingmar Splitt
+author_email = kai.geissdoerfer@tu-dresden.de
+maintainer_email = ingmar.splitt@tu-dresden.de
+license = MIT
+url = https://pypi.org/project/shepherd-cal
+project_urls =
+	Tracker = https://github.com/orgua/shepherd/issues
+	Source = https://github.com/orgua/shepherd
+description = Synchronized Energy Harvesting Emulator and Recorder CLI
+long_description = file: README.md
+long_description_content_type = text/markdown
+keywords = testbed, beaglebone, pru, batteryless, energyharvesting, solar
+platforms = unix, linux, osx, cygwin, win32, win64
+classifiers =
+	Development Status :: 4 - Beta
+	Intended Audience :: Developers
+	Intended Audience :: Information Technology
+	Intended Audience :: Science/Research
+	Programming Language :: Python :: 3
+	Programming Language :: Python :: 3.8
+	Programming Language :: Python :: 3.9
+	Programming Language :: Python :: 3.10
+	License :: OSI Approved :: MIT License
+	Operating System :: OS Independent
+	Natural Language :: English
+
+[options]
+packages = find:
+package_dir =
+	=./
+zip_safe = True
+include_package_data = True
+python_requires = >= 3.8
+install_requires =
+	click
+	matplotlib
+	numpy
+	fabric
+	pyYAML
+	zerorpc
+	msgpack
+	msgpack_numpy
+	shepherd @ git+https://github.com/orgua/shepherd.git#subdirectory=software/python-package/
+	keithley2600 @ git+https://github.com/orgua/keithley2600.git
+
+[options.entry_points]
+console_scripts =
+	shepherd-cal = shepherd_cal.cli:cli
+
+[options.packages.find]
+where = ./
