+++ software/shepherd-calibrate/shepherd_cal/__init__.py
@@ -1,0 +1,13 @@
+from .calibrate import logger
+from .calibrate import set_verbose_level
+from .calibrate import Cal
+from .plot import plot_calibration
+
+__version__ = "0.4.3"
+
+__all__ = [
+    "Cal",
+    "logger",
+    "set_verbose_level",
+    "plot_calibration",
+]
